// apps/host/src/types/hyperliquid.ts
export interface WsActiveAssetCtx {
    coin: string;
    ctx: {
        dayNtlVlm: string;
        dayBaseVlm: string;
        prevDayPx: string;
        markPx: string;
        midPx?: string;
        funding?: number;
        openInterest?: number;
        oraclePx?: number;
        circulatingSupply?: number;
    };
}